requirements.txt
chainlit
google-generativeai
python-dotenv
langchain_groq
langchain
-e .
